draft-coffee-a2a/pyproject.toml
[project]
name = "draft-coffee-mantha"
version = "1.0.0"
description = "MANTHA — General Data Pipeline Agent for the A2A (Agent2Agent) network"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # A2A framework
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "openai>=1.57.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    # MANTHA pipeline dependencies
    "pandas>=2.0",
    "openpyxl>=3.1",
    "xlrd>=2.0",
    "requests>=2.31",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "reportlab>=4.1",
    # Gmail API (falls back to SMTP if not installed)
    "google-auth>=2.28",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.120",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
